refactor(Types): 유지보수 향상을 위한 인터페이스 위치 이동
