+++ src/hooks/useUserName.ts
@@ -3,6 +3,12 @@ import CONDITION from '../constants/Condition';
 
 const { REG_EXP } = CONDITION;
 
+export interface UserNameState {
+  userName: string | undefined;
+  setUserName: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  userNameError: boolean;
+}
+
 const useUserName = (defaultValue: string | undefined) => {
   const [userName, setUserName, userNameError] = useInput(defaultValue, REG_EXP.userName);
