+++ src/types/Types.ts
@@ -1,58 +1,40 @@
-export interface CardNumberState {
-  firstState: FirstState;
-  secondState: SecondState;
-  thirdState: ThirdState;
-  fourthState: FourthState;
-  showImageCondition: ShowImageCondition;
-}
-interface FirstState {
+export interface FirstState {
   first: number | undefined;
   setFirst: (event: React.ChangeEvent<HTMLInputElement>) => void;
   firstError: boolean;
 }
 
-interface SecondState {
+export interface SecondState {
   second: number | undefined;
   setSecond: (event: React.ChangeEvent<HTMLInputElement>) => void;
   secondError: boolean;
 }
 
-interface ThirdState {
+export interface ThirdState {
   third: number | undefined;
   setThird: (event: React.ChangeEvent<HTMLInputElement>) => void;
   thirdError: boolean;
 }
 
-interface FourthState {
+export interface FourthState {
   fourth: number | undefined;
   setFourth: (event: React.ChangeEvent<HTMLInputElement>) => void;
   fourthError: boolean;
 }
 
-interface ShowImageCondition {
+export interface ShowImageCondition {
   isVisa: boolean;
   isMasterCard: boolean;
 }
 
-export interface ExpirationDateState {
-  monthState: MonthState;
-  yearState: YearState;
-}
-
-interface MonthState {
+export interface MonthState {
   month: number | undefined;
   setMonth: (event: React.ChangeEvent<HTMLInputElement>) => void;
   monthError: boolean;
 }
 
-interface YearState {
+export interface YearState {
   year: number | undefined;
   setYear: (event: React.ChangeEvent<HTMLInputElement>) => void;
   yearError: boolean;
-}
-
-export interface UserNameState {
-  userName: string | undefined;
-  setUserName: (event: React.ChangeEvent<HTMLInputElement>) => void;
-  userNameError: boolean;
 }
